emulators/wine-devel: Update to Wine 9.9
This includes the following changes:
 - Removal of a number of obsolete features in WineD3D.
 - Support for new WoW64 mode in ODBC.
 - Various bug fixes.

When trying to build with LLVM 15, which is the current default in
the Ports Collection we have been using for a while, we now get the
following error:

  configure: error: Suitable PE cross-compiler not found, PE files
  won't be built. This is an error since --with-mingw was requested.
  ===>  Script "configure" failed unexpectedly.

This is not actually a regression from Wine 9.8 to 9.9; rather Wine
now diagnoses this (again) instead of simply proceeding.

Luckily our devel/llvm18 and devel/llvm17 ports just recently addressed
the situation with commits c56fde6514 on May 28 and b21e6b4de1 on June 12,
respectively. Accordingly force the use of at least devel/llvm17. [2]

This brings a huge number of changes in terms of pkg-plist.

(The emulators/wine port will need similar changes or it'll break
with builds of LLVM that contain these changes already in the two
ports above. [1])

PR:		279677 [1], 274542 [2]