emulators/wine-devel: Update to Wine 9.13
This includes the following changes:
 - Support for loading ODBC Windows drivers
   (though we haven't enabled ODBC at this point).
 - More user32 data structures in shared memory.
 - More rewriting of the CMD.EXE engine.
 - Various bug fixes.

On the way remove files/clang/stdarg.h, a copy of clang's stdarg.h which
we have been using while the devel/llvm* ports did not install it (and
some others). This changed for devel/llvm17 with commit b21e6b4de1
seven weeks ago.